ContainerServer constructor fixes
Introduced new Paths class to hold path constants. This can
serve as a replacement for the more sophisticated Path
enum since Curator has the ability to ensure full
path creation.

Introduced use of Curator namespace feature to automatically
prefix all paths with /xd.

Moved initialization of PathChildrenCache from ContainerServer
constructor to onConnect. During testing I noticed that
a cache cannot be reopened when closed, so a new instance
is created when a connection is established.

Moved StreamWriter to the tests area since it has a dependency
on a test class.