Bug fixes
When a SUSPENDED event is raised we cancel the leader selector
and the containers childre cache. If a LOST event is the next
event we attempt another close which results in IllegalStateException.
This exception can be ignored since these objects were no longer
in a usable state.

Fixed an NPE in ContainerListener.

Logged leader interruption at level INFO.