Fixed command line arg issue with test
After the migration to Boot we changed the command line arg
format to --zk=host:port. These changes were not reflected
in the tests which caused them to fall back to the default
ZK host:port. I noticed this as a problem after taking down
the ZK server and seeing the tests fail.