Fixed path issue
The leader admin was creating a path of /xd/xd/admin. This
was caused by the unnecessary prepending of /xd to the
Paths.ADMIN string. It turns out that the leadership
API requires that paths contain a leading / whereas
the rest of the Curator client API will add a leading
/ when a namespace is declared upon client construction.

Therefore, in order to be consistent across the board,
the leading / is always provided by Paths.build.